pyproject.toml
[tool.poetry]
name = "online-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dependence.dependencies]
django = "4.2.2"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
